+++ lib/src/components/cards/group_cards/event_group_card.dart
@@ -4,7 +4,6 @@ import 'dart:ui';
 import 'package:dots_design_system/dots_design_system.dart';
 import 'package:figma_squircle/figma_squircle.dart';
 import 'package:flutter/material.dart';
-import 'package:soft_edge_blur/soft_edge_blur.dart';
 
 enum EventGroupCardVariant {
   small,
@@ -95,24 +94,8 @@ class EventGroupCard extends StatelessWidget {
                     cornerRadius: styleType.radius,
                     cornerSmoothing: styleType.cornerSmoothing,
                   ),
-                  child: SoftEdgeBlur(
-                    edges: [
-                      EdgeBlur(
-                        type: EdgeType.bottomEdge,
-                        size: bottomBlurHeight,
-                        sigma: 12,
-                        controlPoints: [
-                          ControlPoint(
-                            position: 0.5,
-                            type: ControlPointType.visible,
-                          ),
-                          ControlPoint(
-                            position: 1,
-                            type: ControlPointType.transparent,
-                          ),
-                        ],
-                      ),
-                    ],
+                  child: DotsBottomEdgeBlur(
+                    edgeSize: bottomBlurHeight,
                     child: Image(
                       image: imageProvider,
                       fit: BoxFit.cover,