+++ lib/src/components/containers/containers_lib.dart
@@ -1,4 +1,5 @@
 export 'dots_main_container/dots_main_container.dart';
 export 'dots_main_container/dots_main_container_enums.dart';
 export 'blur_container.dart';
+export 'dots_bottom_edge_blur.dart';
 export 'dots_linear_gradient_blur.dart';